Fix chromium version, as the latest (at the time of writing) segfaulted on the pi.
Also update partition sizes to make the resulting image smaller.

+++ home/on-pi-startup.sh
@@ -2,10 +2,12 @@
 
 # Make sure Chromium profile is marked clean, even if it crashed
 # https://stedolan.github.io/jq/manual/
-cat .config/chromium/Default/Preferences \
-    | jq '.profile.exit_type = "SessionEnded" | .profile.exited_cleanly = true' \
-    > .config/chromium/Default/Preferences-clean
-mv .config/chromium/Default/Preferences{-clean,}
+if [ -f .config/chromium/Default/Preferences ]; then
+    cat .config/chromium/Default/Preferences \
+        | jq '.profile.exit_type = "SessionEnded" | .profile.exited_cleanly = true' \
+        > .config/chromium/Default/Preferences-clean
+    mv .config/chromium/Default/Preferences{-clean,}
+fi
 
 # http://peter.sh/experiments/chromium-command-line-switches/
 chromium-browser --start-fullscreen
